refactor: general refactoring (#23)
* feat: add SIGTERM handler

* fix: change log level to INFO for current clock updates

* fix: handle message queue errors and improve process exit handling

* fix: replace exit code 1 with EXIT_FAILURE for memory allocation errors

* fix: cast sumWait to double for accurate average waiting time calculation

* fix: clean up message queue and terminate process group on exit

* fix: increment index in parseCommandLineArgs for valid filename handling

* fix: detaching shm once

* fix: destroy msgqueue in pg only

* fix: pg waits for child termination on sigchild

* fix: proper include for signal.h

* fix: clearResources on exit

* fix: fix import killpg

* refcator: clock period 1 second

---------

+++ src/utils/logger.c
@@ -46,7 +46,7 @@ void logSchedulerPerformance(int idleTime, int totalTime, int numProcesses, doub
 
     // Calculate average waiting time and WTA
     double avgWTA = sumWTA / numProcesses;
-    double avgWait = sumWait / numProcesses;
+    double avgWait = (double)sumWait / numProcesses;
 
     // Calculate standard deviation of WTA
     double varianceWTA = (sumWTAsquared / numProcesses) - (avgWTA * avgWTA);
